Added check if file extists before attempting to delete.
Former-commit-id: 1de513c047e680a3bc604da1313ba3ce9886f4f1

+++ packages/photon/src/generation/generateClient.ts
@@ -140,7 +140,9 @@ export async function generateClient({
     const filePath = path.join(outputDir, fileName)
     // The deletion of the file is necessary, so VSCode
     // picks up the changes.
-    await fs.unlink(filePath)
+    if (fs.exists(filePath)) {
+      await fs.unlink(filePath)
+    }
     await fs.writeFile(filePath, file)
   }))
   await fs.copy(path.join(__dirname, '../../runtime'), path.join(outputDir, '/runtime'))
